requirements.txt
asgiref==3.9.2
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
mysqlclient==2.2.7
sqlparse==0.5.3
